+++ packages/mds-transaction-service/service/provider.ts
@@ -15,8 +15,7 @@
  */
 
 import { ProcessController, ServiceException, ServiceProvider, ServiceResult } from '@mds-core/mds-service-helpers'
-import { UUID } from '@mds-core/mds-types'
-import { TransactionSearchParams, TransactionService } from '../@types'
+import { TransactionService, TransactionServiceRequestContext } from '../@types'
 import { TransactionServiceLogger } from '../logger'
 import { TransactionRepository } from '../repository'
 import { TransactionStreamKafka } from './stream'
@@ -28,14 +27,15 @@ import {
   validateTransactionStatusDomainModel
 } from './validators'
 
-export const TransactionServiceProvider: ServiceProvider<TransactionService> & ProcessController = {
+export const TransactionServiceProvider: ServiceProvider<TransactionService, TransactionServiceRequestContext> &
+  ProcessController = {
   start: async () => {
     await Promise.all([TransactionRepository.initialize(), TransactionStreamKafka.initialize()])
   },
   stop: async () => {
     await Promise.all([TransactionRepository.shutdown(), TransactionStreamKafka.shutdown()])
   },
-  createTransaction: async transaction => {
+  createTransaction: async (context, transaction) => {
     try {
       const recordedTransaction = await TransactionRepository.createTransaction(
         validateTransactionDomainModel(transaction),
@@ -52,7 +52,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
       return exception
     }
   },
-  createTransactions: async transactions => {
+  createTransactions: async (context, transactions) => {
     try {
       const recordedTransactions = await TransactionRepository.createTransactions(
         transactions.map(validateTransactionDomainModel),
@@ -69,7 +69,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
       return exception
     }
   },
-  getTransaction: async (transaction_id: UUID) => {
+  getTransaction: async (context, transaction_id) => {
     try {
       const transaction = await TransactionRepository.getTransaction(transaction_id)
       return ServiceResult(transaction)
@@ -80,7 +80,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
     }
   },
   // TODO search params
-  getTransactions: async (search: TransactionSearchParams) => {
+  getTransactions: async (context, search) => {
     try {
       const transactions = await TransactionRepository.getTransactions(search)
       return ServiceResult(transactions)
@@ -90,7 +90,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
       return exception
     }
   },
-  addTransactionOperation: async transactionOperation => {
+  addTransactionOperation: async (context, transactionOperation) => {
     try {
       const operation = await TransactionRepository.addTransactionOperation(
         validateTransactionOperationDomainModel(transactionOperation)
@@ -103,7 +103,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
     }
   },
   // TODO search params
-  getTransactionOperations: async (transaction_id: UUID) => {
+  getTransactionOperations: async (context, transaction_id) => {
     try {
       const operations = await TransactionRepository.getTransactionOperations(transaction_id)
       return ServiceResult(operations)
@@ -113,7 +113,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
       return exception
     }
   },
-  setTransactionStatus: async transactionStatus => {
+  setTransactionStatus: async (context, transactionStatus) => {
     try {
       const status = await TransactionRepository.setTransactionStatus(
         validateTransactionStatusDomainModel(transactionStatus)
@@ -126,7 +126,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
     }
   },
   // TODO search params
-  getTransactionStatuses: async (transaction_id: UUID) => {
+  getTransactionStatuses: async (context, transaction_id) => {
     try {
       validateTransactionId(transaction_id)
       const statuses = await TransactionRepository.getTransactionStatuses(transaction_id)
@@ -137,7 +137,7 @@ export const TransactionServiceProvider: ServiceProvider<TransactionService> & P
       return exception
     }
   },
-  getTransactionsStatuses: async (transaction_ids: UUID[]) => {
+  getTransactionsStatuses: async (context, transaction_ids) => {
     try {
       validateTransactionIds(transaction_ids)
       const statuses = await TransactionRepository.getTransactionsStatuses(transaction_ids)