+++ container-images/mds-jurisdiction-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-jurisdiction-api
 
+## 0.0.41
+
+### Patch Changes
+
+- @mds-core/mds-jurisdiction-api@0.1.40
+
 ## 0.0.40
 
 ### Patch Changes