+++ packages/mds-agency/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mds-core/mds-agency
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [c3d2215d]
+  - @mds-core/mds-ingest-service@0.8.0
+  - @mds-core/mds-service-helpers@0.5.0
+  - @mds-core/mds-db@0.4.26
+  - @mds-core/mds-api-server@0.3.20
+  - @mds-core/mds-api-helpers@0.1.66
+
 ## 0.4.2
 
 ### Patch Changes