+++ packages/mds-collector-api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mds-core/mds-collector-api
 
+## 0.1.41
+
+### Patch Changes
+
+- Updated dependencies [c3d2215d]
+  - @mds-core/mds-collector-service@0.4.0
+  - @mds-core/mds-service-helpers@0.5.0
+  - @mds-core/mds-api-server@0.3.20
+  - @mds-core/mds-api-helpers@0.1.66
+
 ## 0.1.40
 
 ### Patch Changes