+++ packages/mds-stream-processor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mds-core/mds-stream-processor
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies [c3d2215d]
+  - @mds-core/mds-service-helpers@0.5.0
+
 ## 0.4.3
 
 ### Patch Changes